remember-me-server/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-jose
python-multipart
passlib
pydantic
python-dotenv
pydantic_settings
sqlalchemy
pydantic[email]
image
psycopg2-binary
alembic
chromadb
langchain
langchain-chroma
langchain-community
openai
langchain-openai
langchain-google-genai
google-generativeai
langchain-anthropic